Allow kernel to free pages of memory from arenas
Upon resetting an arena to a previous point, be it the beginning of the
buffer or a checkpoint, use `madvise()` with `MADV_FREE` to advise (not
command) the kernel that it's unlikely the program needs the range of
memory outside the highest allocated page bound in the buffer of the
arena allocator in the near future. In which case, the kernel may free
those unused pages if memory pressure occurs and then provide them
again, without another mmap(), for new writes in the range.