Only update title if feed already exists in DB
Use `UPSERT` instead of `REPLACE` clause to insert a new feed or update
the title of an existing feed.

With `REPLACE`, SQLite deletes and then again inserts a feed. When the
storage engine deletes a feed, it also deletes all of the feeds
corresponding items because of the `CASCADE` constraint, effectively
losing the state of each item's read/unread attribute, which is not
desirable.

`UPSERT` fixes this behavior because the storage engine no longer
deletes entries that violate the implicit unique constraint of the
primary key.